requirements.txt
absl-py==1.3.0
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
black==22.3.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
contourpy==1.0.6
cycler==0.11.0
detectron2==0.6
fairscale==0.4.12
filelock==3.8.0
Flask==2.2.2
Flask-RESTful==0.3.9
fonttools==4.38.0
future==0.18.2
fvcore==0.1.5.post20220512
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-aiplatform==1.18.3
google-cloud-bigquery==2.34.4
google-cloud-core==2.3.2
google-cloud-resource-manager==1.6.3
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-status==1.48.2
httplib2==0.21.0
huggingface-hub==0.11.0
hydra-core==1.2.0
idna==3.4
importlib-metadata==5.0.0
iopath==0.1.9
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mypy-extensions==0.4.3
numpy==1.23.4
oauthlib==3.2.2
omegaconf==2.2.3
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.1
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.5.4
portalocker==2.6.0
proto-plus==1.22.1
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pydot==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.0
threadpoolctl==3.1.0
timm==0.6.11
tomli==2.0.1
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
yacs==0.1.8
zipp==3.10.0